REO progress - binary logs

+++ java_console/ui/src/main/java/com/rusefi/ui/FormulasPane.java
@@ -210,7 +210,7 @@ public class FormulasPane {
         double displacement = ConfigField.getFloatValue(ci, Fields.DISPLACEMENT);
         int cylinderCount = ConfigField.getIntValue(ci, Fields.CYLINDERSCOUNT);
         String cylinderDisplacement = oneDecimal(displacement / cylinderCount);
-        String injectorFlow = oneDecimal((float) Fields.INJECTOR_FLOW.getValue(ci));
+        String injectorFlow = oneDecimal((float) Fields.INJECTOR_FLOW.getValue(ci).doubleValue());
 
         String tCharge = "$Tcharge=f(CLT=" + oneDecimal(Sensor.CLT) + "C,IAT=" + IAT
                 + "C,TPS=" + tpsStr + "\\%, RPM = " + RPM + ")=" + T_CHARGE + "C$";
